feat: create users table and it can be seen in postgres table.

+++ backend-kotlin-development/docker/db/sql/create_tables.sql
@@ -1,5 +1,5 @@
 CREATE TABLE users (
-  id bigint NOT NULL,
+  id SERIAL NOT NULL,
   name varchar(255) NOT NULL,
   created_at timestamp NOT NULL,
   updated_at timestamp NOT NULL,
@@ -7,7 +7,7 @@ CREATE TABLE users (
 );
 
 CREATE TABLE portfolios (
-  id bigint NOT NULL,
+  id SERIAL NOT NULL,
   user_id bigint NOT NULL,
   title varchar(255) NOT NULL,
   created_at timestamp,
@@ -17,7 +17,7 @@ CREATE TABLE portfolios (
 );
 
 CREATE TABLE brands (
-  id bigint NOT NULL,
+  id SERIAL NOT NULL,
   name varchar(255) NOT NULL,
   price integer NOT NULL,
   quantity integer,
@@ -30,7 +30,7 @@ CREATE TABLE brands (
 );
 
 CREATE TABLE portfolios_belong_to_brand (
-  id bigint NOT NULL,
+  id SERIAL NOT NULL,
   portfolio_id bigint NOT NULL,
   brand_id bigint NOT NULL,
   created_at timestamp,
@@ -41,7 +41,7 @@ CREATE TABLE portfolios_belong_to_brand (
 );
 
 CREATE TABLE strategies (
-  id bigint NOT NULL,
+  id SERIAL NOT NULL,
   brand_id bigint NOT NULL,
   name varchar(255) NOT NULL,
   created_at timestamp,
